[[ Win64 ]] Update engine to compile for win32 x86-64
This patch tweaks the engine code to compile for win32 x86-64.

Mainly it uses intptr_t in the gentle related toolchain code
rather than long, and DWORD_PTR / LONG_PTR appropriately in
Win32 APIs.

Additionally, it updates the build-libs.bat script to use v140
toolset (even if VS 15 is requested - this is because VS2017 does
not have the same structure, and it only bumps the toolset version
by 1, not 10 - i.e. v141 and not v150).

Finally, it modifies fetch-libraries.sh to ensure that the win32
prebuilts have the correct names. The system expects the toolset
to be v141, and for 64-bit it expects the arch to be x64.