[[ VS2015 ]] Restructure Windows prebuilts
This patch reworks both the build files and names of windows
prebuilt libraries for compatibility with VS2015.

The main change is that the type and version of CRT used to build
the library is now part of the name of the prebuilt. This is due
to the fact that VS2015 will not allow you to link together static
libraries which use a mixture of Debug and Release CRTs.

Each prebuilt library now installs into its own folder under
the prebuilt/unpacked folder. The relevant gyp files in this
repo and in thirdparty have been updated to reflect that for
windows only.

The fetch-libraries script has had windows-specific actions added
including a hard-coded exception for CEF which (as it contains no
static libraries) has no dependence on CRT or visual studio version.

The build-libs batch script can be called with a single parameter
to indicate one of openssl, curl or icu to be built. The build uses
the ARCH, MODE and TOOL environment variables to determine the build.
For example ARCH=x86, MODE=debug and TOOL=14 will build a prebuilt
using the VS2015 toolchain (which is v140), the static debug CRT (/MTd
option) for x86.

There is a new batch script build-all-libs which builds four variants
of a library simultaneously - debug/x86, debug/x86_64, release/x86 and
release/x86_64. It takes a single parameter indicating which library
to build.

Note: ICU does not like the parallelization used for build-all-libs so
ICU must be built individually using build-libs and setting env vars.