[[ Bug 19658 ]] Make CEF work on Linux
This patch further changes the layout of CEF components on Linux.

The -bin folder generated from building on Linux now reflects the
same structure as required at runtime:

   engine
   *-cefprocess
   icu.dat
   *.bin
   Externals/CEF/libcef.so
   Externals/CEF/*

This is also reflected in the installer layout, both for 'Runtime'
components, and the set of externals for the IDE as well as what
is generated by the standalone builder.