[[ Build ]] Only fetch prebuilts for target arch
Previously all arguments to the fetch-libraries.sh script were
interpreted as platforms. This patch actually checks the arguments
are valid platforms, and if not, assumes only the first argument
is a platform and subsequent args are target architectures.

The prebuilt fetch gyp action has been modified to call the script
with the appropriate platform and architecture arguments.